move docs cache purging to its own management command
